backend/requirements.txt
attrs==21.2.0
falcon==3.0.1
gunicorn==20.1.0
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.18.0
PyYAML==5.4.1
toml==0.10.2
